requirements.txt
streamlit==1.32.2
openai==0.28.1
python-dotenv
